requirements.txt
elasticsearch==8.14.0
Flask==3.1.0
langchain==0.3.7
langchain_openai==0.2.8
langchain_text_splitters==0.3.2
llmsherpa==0.1.4
marker-pdf==0.2.17
numpy
openai==1.54.4
pandas
Pillow
pypdfium2==4.30.0
python-dotenv==1.0.1
ragas==0.2.5
Spire.Pdf==10.8.1
torch==2.3.1
transformers==4.45.0
flash-attn==2.7.0.post2
sentence-transformers==3.3.0
rouge_score==0.1.2
python_Levenshtein==0.26.1
PyYAML==6.0.2
nltk==3.9.1
tokenizers==0.20